+++ playgrounds/skeleton-react/src/routes/components/tooltip/index.tsx
@@ -15,12 +15,7 @@ function Page() {
 				<Tooltip.Positioner>
 					<Tooltip.Content>
 						Content
-						<Tooltip.Arrow
-							style={{
-								['--arrow-size' as string]: 'calc(var(--spacing) * 2)',
-								['--arrow-background' as string]: 'var(--color-surface-100-900)',
-							}}
-						>
+						<Tooltip.Arrow className="[--arrow-size:--spacing(2)] [--arrow-background:var(--color-surface-100-900)]">
 							<Tooltip.ArrowTip />
 						</Tooltip.Arrow>
 					</Tooltip.Content>