+++ sites/skeleton.dev/src/components/examples/framework-components/tooltip/svelte/arrow.svelte
@@ -6,9 +6,9 @@
 	<Tooltip.Trigger>Hover</Tooltip.Trigger>
 	<Portal>
 		<Tooltip.Positioner>
-			<Tooltip.Content class="card bg-surface-100-900 p-2  shadow-xl">
+			<Tooltip.Content class="card bg-surface-100-900 p-2 shadow-xl">
 				<span>Hello Skeleton</span>
-				<Tooltip.Arrow style="--arrow-size: calc(var(--spacing) * 2); --arrow-background: var(--color-surface-100-900);">
+				<Tooltip.Arrow class="[--arrow-size:--spacing(2)] [--arrow-background:var(--color-surface-100-900)]">
 					<Tooltip.ArrowTip />
 				</Tooltip.Arrow>
 			</Tooltip.Content>